Creating quiz update (endpoint) v2

+++ music-api/src/main/java/com/music/quiz/model/QuizRequest.java
@@ -3,5 +3,5 @@ package com.music.quiz.model;
 public record QuizRequest(
         Long albumId,
         Long artistId,
-        Integer numberOfWords) {
+        Long numberOfWords) {
 }
