+++ music-library/src/main/java/com/music/quiz/dto/QuizRequestDto.java
@@ -3,6 +3,6 @@ package com.music.quiz.dto;
 public record QuizRequestDto(
         Long albumId,
         Long artistId,
-        Integer numberOfWords) {
+        Long numberOfWords) {
 }
